FIT-80: Implement delete function with popup menu
