--- src/features/workout-tracker/components/WorkoutHistoryCard.tsx
@@ -1,38 +1,0 @@
-import React from 'react';
-import { FlatList } from 'react-native';
-
-import { Container, Spacer, Text } from '../../../components';
-import { type Workout } from '../../../interfaces/Workout';
-import { pluralizeWord } from '../../../utils/PluarlizeWord';
-
-interface Props {
-    workout: Workout;
-}
-
-export default function WorkoutHistoryCard({ workout }: Props): React.ReactElement {
-    const date = new Date(workout.dateCreated);
-    return (
-        <Container
-            onPress={() => {
-                // TODO: Open workout history modal
-            }}
-        >
-            <Text variant='headline'>{workout.name}</Text>
-            <Text variant='subhead' color='light'>
-                {date.toDateString()}
-            </Text>
-            <Spacer size='xxs' />
-            <Text variant='headline'>Exercises</Text>
-            <FlatList
-                data={workout.exercises}
-                renderItem={({ item }) => (
-                    <Text variant='body'>
-                        {item.name?.name}: {item.sets.length}{' '}
-                        {pluralizeWord(item.sets.length, 'set')}
-                    </Text>
-                )}
-                keyExtractor={(item, index) => index.toString()}
-            />
-        </Container>
-    );
-}
