+++ src/features/workout-tracker/components/WorkoutHistoryCard/Styles.ts
@@ -1,0 +1,10 @@
+import { View } from 'react-native';
+import styled from 'styled-components';
+
+export const Title = styled(View)``;
+
+export const Header = styled(View)`
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+`;
